Reduce scope of classify response
