+++ src/core/run_queries.hpp
@@ -1,6 +1,8 @@
 #pragma once
 
 #include "configs.hpp"
+#include "query_classify.hpp"
+#include "query_edit_code.hpp"
 #include "responses_ollama.hpp"
 #include "responses_openai.hpp"
