+++ src/queries/responses_ollama.cpp
@@ -27,22 +27,6 @@ OllamaResponse::OllamaResponse(const std::string &response) :
     }
 }
 
-OllamaClassification::OllamaClassification(const std::string &response) :
-    OllamaResponse(response)
-{
-    nlohmann::json structured_output;
-
-    try {
-        const std::string content = this->response_["message"]["content"];
-        structured_output = nlohmann::json::parse(content);
-    } catch (const nlohmann::json::parse_error &e) {
-        throw std::runtime_error(fmt::format("Failed to parse structured output: {}", e.what()));
-    }
-
-    this->valid_instructions = structured_output.at("valid_instructions").get<bool>();
-    this->reasoning = structured_output["reasoning"];
-}
-
 OllamaEdit::OllamaEdit(const std::string &response, const double total_t) :
     OllamaResponse(response), total_time(total_t)
 {
